Empty RecordSets are falsy now
RecordSets have the quirk that they're actually bundles of bunches of
records. A RecordSet may contain one empty group, but that'll erroneously
look truthy because there actually is a group. But we typically want
to know if it's empty.

I'm not changing it's behavior, though. RecordSets have this as a core
feature because they allow for and enable immutable delta updates.